# pulsar-rest/src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: Browser4 WebDriver-Compatible API
  description: |
    A WebDriver-compatible HTTP API with selector-first extensions for browser automation.
    This API provides session management, navigation, element interaction, and event handling
    capabilities.
  version: 1.0.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8182
    description: Local development server
...
paths:
  /session:
    post:
      tags:
        - session
      summary: Create a new browser session
      operationId: createSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewSessionRequest'
      responses:
        '200':
          description: Session created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewSessionResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}:
    get:
      tags:
        - session
      summary: Get session details
      operationId: getSession
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDetails'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - session
      summary: Delete a session
      operationId: deleteSession
      parameters:
        - $ref: '#/components/parameters/sessionId'
      responses:
        '200':
          description: Session deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebDriverResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /session/{sessionId}/selectors/exists:
    post:
      tags:
        - selectors
      summary: Check if selector exists
      operationId: selectorExists
      parameters:
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SelectorRef'
      responses:
        '200':
          description: Existence check result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExistsResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/selectors/waitFor:
    post:
      tags:
        - selectors
      summary: Wait for selector to appear
      operationId: waitForSelector
      parameters:
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WaitForRequest'
      responses:
        '200':
          description: Selector found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebDriverResponse'
        '408':
          description: Timeout waiting for selector
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/selectors/element:
    post:
      tags:
        - selectors
      summary: Find element by selector
      operationId: findElementBySelector
      parameters:
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SelectorRef'
      responses:
        '200':
          description: Element found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ElementResponse'
        '404':
          description: Element not found or session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/selectors/elements:
    post:
      tags:
        - selectors
      summary: Find all elements matching selector
      operationId: findElementsBySelector
      parameters:
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SelectorRef'
      responses:
        '200':
          description: Elements found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ElementsResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/selectors/click:
    post:
      tags:
        - selectors
      summary: Click element by selector
      operationId: clickBySelector
      parameters:
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SelectorRef'
      responses:
        '200':
          description: Click performed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebDriverResponse'
        '404':
          description: Element or session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/selectors/fill:
    post:
      tags:
        - selectors
      summary: Fill input element by selector
      operationId: fillBySelector
      parameters:
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FillRequest'
      responses:
        '200':
          description: Input filled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebDriverResponse'
        '404':
          description: Element or session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/selectors/press:
    post:
      tags:
        - selectors
      summary: Press key on element by selector
      operationId: pressBySelector
      parameters:
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PressRequest'
      responses:
        '200':
          description: Key pressed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebDriverResponse'
        '404':
          description: Element or session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/selectors/outerHtml:
    post:
      tags:
        - selectors
      summary: Get outer HTML of element by selector
      operationId: getOuterHtmlBySelector
      parameters:
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SelectorRef'
      responses:
        '200':
          description: Outer HTML content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HtmlResponse'
        '404':
          description: Element or session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /session/{sessionId}/selectors/screenshot:
    post:
      tags:
        - selectors
      summary: Take screenshot of element by selector
      operationId: screenshotBySelector
      parameters:
        - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SelectorRef'
      responses:
        '200':
          description: Screenshot captured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScreenshotResponse'
        '404':
          description: Element or session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  parameters:
    sessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
      description: Session identifier
...
  schemas:
    NewSessionRequest:
      type: object
      properties:
        capabilities:
          type: object
          description: Desired capabilities for the session
          additionalProperties: true
    NewSessionResponse:
      type: object
      properties:
        value:
          type: object
          properties:
            sessionId:
              type: string
            capabilities:
              type: object
              additionalProperties: true
    SessionDetails:
      type: object
      properties:
        value:
          type: object
          properties:
            sessionId:
              type: string
            url:
              type: string
            status:
              type: string
              enum: [active, paused, stopped]
            capabilities:
              type: object
              additionalProperties: true
    WebDriverResponse:
      type: object
      properties:
        value:
          description: Response value (can be null for success)
    ErrorResponse:
      type: object
      properties:
        value:
          type: object
          properties:
            error:
              type: string
            message:
              type: string
            stacktrace:
              type: string
...
    SelectorRef:
      type: object
      required:
        - selector
      properties:
        selector:
          type: string
          description: CSS selector or XPath expression
        strategy:
          type: string
          enum: [css, xpath]
          default: css
          description: Selector strategy
    WaitForRequest:
      type: object
      required:
        - selector
      properties:
        selector:
          type: string
        strategy:
          type: string
          enum: [css, xpath]
          default: css
        timeout:
          type: integer
          default: 30000
          description: Timeout in milliseconds
...
    ElementRef:
      type: object
      properties:
        element-6066-11e4-a52e-4f735466cecf:
          type: string
          description: WebDriver element identifier
    ElementResponse:
      type: object
      properties:
        value:
          $ref: '#/components/schemas/ElementRef'
    ElementsResponse:
      type: object
      properties:
        value:
          type: array
          items:
            $ref: '#/components/schemas/ElementRef'
    ExistsResponse:
      type: object
      properties:
        value:
          type: object
          properties:
            exists:
              type: boolean
    FillRequest:
      type: object
      required:
        - selector
        - value
      properties:
        selector:
          type: string
        strategy:
          type: string
          enum: [css, xpath]
          default: css
        value:
          type: string
          description: Text to fill
    PressRequest:
      type: object
      required:
        - selector
        - key
      properties:
        selector:
          type: string
        strategy:
          type: string
          enum: [css, xpath]
          default: css
        key:
          type: string
          description: Key to press (e.g., Enter, Tab)
    HtmlResponse:
      type: object
      properties:
        value:
          type: string
          description: HTML content
    ScreenshotResponse:
      type: object
      properties:
        value:
          type: string
          format: byte
          description: Base64-encoded PNG screenshot
...